Add View fixes and minor bug fixes

+++ LearnEase/src/LearnEase.Infrastructure/Services/LessonService.cs
@@ -9,11 +9,14 @@ public class LessonService : ILessonService
 {
     private readonly ILessonRepository lessonRepository;
 
+    private readonly ICourseRepository courseRepository;
+
     private readonly UserManager<User> userManager;
 
-    public LessonService(ILessonRepository lessonRepository, UserManager<User> userManager) {
+    public LessonService(ILessonRepository lessonRepository, ICourseRepository courseRepository, UserManager<User> userManager) {
         this.lessonRepository = lessonRepository;
         this.userManager = userManager;
+        this.courseRepository = courseRepository;
     }
 
     public async Task<Lesson> GetLessonByIdAsync(int lessonId)
@@ -30,6 +33,12 @@ public class LessonService : ILessonService
     {
         newLesson.AuthorName = user.UserName;
         newLesson.Timestamp = DateTime.Now;
+        var course = await courseRepository.GetByIdAsync(newLesson.CourseId);
+
+        if (course is null)
+            throw new ArgumentNullException("Course is null!");
+
+        course.AmountOfLectures += 1;
 
         var changesCount = await lessonRepository.CreateAsync(newLesson);
 
@@ -39,7 +48,17 @@ public class LessonService : ILessonService
 
     public async Task DeleteLessonByIdAsync(int lessonId)
     {
+        var lesson = await lessonRepository.GetByIdAsync(lessonId);
+        var course = await courseRepository.GetByIdAsync(lesson.CourseId);
+
+        if (course is null)
+            throw new ArgumentNullException("Course is null!");
+
+        course.AmountOfLectures -= 1;
+
         var changesCount = await lessonRepository.DeleteAsync(lessonId);
+
+
 
         if (changesCount == 0)
             throw new Exception("Lesson delete didn't apply!");
